requirements.txt
streamlit == 1.30
numpy == 1.26
msgpack == 1.0.7
pillow == 10.2
pandas == 2.1.4

